+++ src/db/migrations/20190408115030-add-validation-columns-users.js
@@ -1,0 +1,26 @@
+module.exports = {
+    up: async (queryInterface, Sequelize) => {
+        await queryInterface.addColumn('Users', 'contactMethodVerificationToken', {
+            type: Sequelize.STRING,
+            unique: true,
+        });
+        await queryInterface.addColumn('Users', 'contactMethodVerificationTokenExpires', {
+            type: Sequelize.DATE,
+        });
+        await queryInterface.addColumn('Users', 'contactMethodToVerify', {
+            type: Sequelize.STRING, // eslint-disable-line new-cap
+            defaultValue: '',
+        });
+        await queryInterface.addColumn('Users', 'verifiedContactMethods', {
+            type: Sequelize.ARRAY(Sequelize.STRING), // eslint-disable-line new-cap
+            defaultValue: [],
+        });
+        await queryInterface.sequelize.query(`UPDATE "Users" SET "verifiedContactMethods" = ARRAY["email"]`)
+    },
+    down: async (queryInterface) => {
+        await queryInterface.removeColumn('Users', 'contactMethodVerificationToken');
+        await queryInterface.removeColumn('Users', 'contactMethodVerificationTokenExpires');
+        await queryInterface.removeColumn('Users', 'contactMethodToVerify');
+        await queryInterface.removeColumn('Users', 'verifiedContactMethods');
+    },
+};