+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Added
 - ENV `ALWAYS_INCLUDE_ERROR_STACKS` to include full error stack, including any causal errors.
   * Use with caution in production, because the full stack could leak sensitive information.
+- Optional email verification for users
+  * When `AUTH_SERVICE_REQUIRE_ACCOUNT_VERIFICATION=true`, a user cannot sign-in without email verification
+  * When `AUTH_SERVICE_REQUIRE_SECURE_ACCOUNT_VERIFICATION=true`, a user password is required for email verification
 
 ### Fixed
 - `auth.controller.js:login()` no longer throws duplicate errors for incorrect username or password